Squashed commit of the following:
commit 5102aefdf3e1710f3ca9bc00c3e36de93864a59b
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 19:02:15 2019 +0530

    Refactored code

commit 374e19cda25816371a05c1b63b22f6fecb066e4b
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 18:54:04 2019 +0530

    Updated submit method

commit 3fdbcf9bb0e6235b6463096de2e470ad9e560bf7
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 16:41:52 2019 +0530

    Code cleanup

commit e0adbc6c7188c3f832fda51854950f060f323741
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 16:24:23 2019 +0530

    Updated addon-docs version

commit 91bba0a214149b7b78e2af32674351232ca43f44
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 16:05:00 2019 +0530

    Refactored example doc structure

commit 86de3a470e6bf55957a986f82884b94dbc790f69
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 14:36:54 2019 +0530

    Adding loading animations

commit 3dc5d001e3d0a602cbba50a9820258e655f4fa15
Author: shibulijack <shibulijack@SHIBULIJACK-2.local>
Date:   Tue Oct 1 10:54:49 2019 +0530

    Added loading button demo
