+++ addon/constants/nucleus-button.js
@@ -1,0 +1,8 @@
+const BUTTON_STATE = {
+  DEFAULT: "default",
+  PENDING: "pending",
+  FULFILLED: "fulfilled",
+  REJECTED: "rejected"
+};
+
+export { BUTTON_STATE };
