+++ addon/styles/components/_nucleus-button.scss
@@ -29,9 +29,11 @@ $active-box-shadow: inset 0 0 4px 0 rgba(0, 0, 0, .25);
   border-radius: 4px;
   outline: 0;
   color: $btn-primary-color;
+  cursor: pointer;
   transition: all .2s linear;
   letter-spacing: .3px;
   font-weight: $font-weight-500;
+  min-width: 5rem;
 
   &:focus {
     border: 1px solid $color-azure-800;
@@ -114,6 +116,7 @@ $active-box-shadow: inset 0 0 4px 0 rgba(0, 0, 0, .25);
 .btn--mini {
   padding: 4px 5px;
   font-size: 12px;
+  min-width: 1rem;
 
   svg.icon {
     width: 12px;
