+++ addon/styles/global/_animations.scss
@@ -59,3 +59,95 @@
     transition: all 300ms ease-in-out;
   }
 }
+
+//=== Loading button animation ===//
+$loader-size: 1rem;
+$check-height: $loader-size/1.6;
+$check-width: $check-height/2;
+$check-left: $loader-size/4;
+$check-top: $loader-size/2;
+$check-thickness: 2px;
+$check-color: $color-milk;
+
+.circle-loader {
+  border: 2px solid rgba(0, 0, 0, .15);
+  border-left-color: $check-color;
+  animation: loader-spin 1.2s infinite linear;
+  position: relative;
+  display: inline-block;
+  vertical-align: top;
+  border-radius: 50%;
+  width: $loader-size;
+  height: $loader-size;
+}
+
+.load-complete {
+  -webkit-animation: none;
+  animation: none;
+  border: none;
+  transition: all 500ms ease-out;
+
+  .checkmark.draw {
+    display: block;
+  }
+}
+
+.checkmark {
+  display: none;
+
+  &.draw::after {
+    animation-duration: 800ms;
+    animation-timing-function: ease;
+    animation-name: checkmark;
+    transform: scaleX(-1) rotate(135deg);
+  }
+
+  &::after {
+    opacity: 1;
+    height: $check-height;
+    width: $check-width;
+    transform-origin: left top;
+    border-right: $check-thickness solid $check-color;
+    border-top: $check-thickness solid $check-color;
+    content: '';
+    left: $check-left;
+    top: $check-top;
+    position: absolute;
+  }
+}
+
+@keyframes loader-spin {
+  0% {
+    transform: rotate(0deg);
+  }
+
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+@keyframes checkmark {
+  0% {
+    height: 0;
+    width: 0;
+    opacity: 1;
+  }
+
+  20% {
+    height: 0;
+    width: $check-width;
+    opacity: 1;
+  }
+
+  40% {
+    height: $check-height;
+    width: $check-width;
+    opacity: 1;
+  }
+
+  100% {
+    height: $check-height;
+    width: $check-width;
+    opacity: 1;
+  }
+}
