+++ addon/templates/components/nucleus-button.hbs
@@ -1,5 +1,11 @@
 {{#if icon}}{{svg-jar icon class="icon" width="14px"}}{{/if}}
-{{_buttonText}}
+{{#if (and _isLoading _isShowLoading)}}
+  <div data-test-button-loader class="circle-loader {{if _isLoadingComplete "load-complete"}}">
+    <div class="checkmark draw"></div>
+  </div>
+{{else}}
+  {{_buttonText}}
+{{/if}}
 {{#if hasBlock}}
   {{yield}}
 {{/if}}
