+++ app/styles/doc/_custom-addon-theme.scss
@@ -36,3 +36,7 @@
 .docs-md ul {
   font-size: .875rem;
 }
+
+.docs-no-padding {
+  padding: 0 !important; // to prevent blank spaces while rendering logic templates
+}
