+++ package.json
@@ -35,7 +35,7 @@
     "broccoli-asset-rev": "^2.7.0",
     "ember-a11y-testing": "^1.0.0",
     "ember-cli": "~3.7.1",
-    "ember-cli-addon-docs": "^0.6.4",
+    "ember-cli-addon-docs": "^0.6.14",
     "ember-cli-addon-docs-yuidoc": "^0.2.1",
     "ember-cli-code-coverage": "^0.4.2",
     "ember-cli-dependency-checker": "^3.0.0",
@@ -62,6 +62,7 @@
     "ember-sinon-qunit": "^3.4.0",
     "ember-source": "~3.7.0",
     "ember-source-channel-url": "^1.1.0",
+    "ember-test-selectors": "^2.1.0",
     "ember-try": "^1.0.0",
     "eslint-plugin-ember": "^5.2.0",
     "eslint-plugin-node": "^7.0.1",
