+++ tests/dummy/app/components/nucleus-button/demo-1.js
@@ -1,0 +1,17 @@
+// BEGIN-SNIPPET button-loading.js
+import Component from '@ember/component';
+import RSVP from 'rsvp';
+
+export default Component.extend({
+  actions: {
+    download(value) {
+      return new RSVP.Promise((resolve) => {
+        let wait = setTimeout(() => {
+          clearTimeout(wait);
+          resolve(value);
+        }, 1000)
+      });
+    }
+  }
+});
+// END-SNIPPET
