+++ tests/dummy/app/components/nucleus-modal/demo-2.js
@@ -1,0 +1,11 @@
+// BEGIN-SNIPPET dialog-component.js
+import Component from '@ember/component';
+
+export default Component.extend({
+  actions: {
+    onSubmit(value) {
+      alert(value);
+    }
+  }
+});
+// END-SNIPPET
