+++ tests/dummy/app/templates/components/nucleus-modal/demo-2.hbs
@@ -1,0 +1,20 @@
+{{#docs-demo as |demo|}}
+  {{#demo.example name="modal-demo-2-helper.hbs"}}
+    <button onclick={{action (mut isModal2 true)}} class="docs-btn">
+      Open confirmation dialog
+    </button>
+  {{/demo.example}}
+  {{#demo.example name="modal-demo-2.hbs" class="docs-no-padding"}}
+    {{#nucleus-confirm-dialog
+      open=isModal2
+      isDismissible=false
+      type="danger"
+      onSubmit=(action "onSubmit" "Submit action called")
+      title="Account cancellation"}}
+      <p>Your account will be cancelled in <b>7 days</b> on <b>July 4, 2019</b>. Your account and data will be irrevocably deleted two weeks later.</p>
+    {{/nucleus-confirm-dialog}}
+  {{/demo.example}}
+
+  {{demo.snippet "modal-demo-2.hbs"}}
+  {{demo.snippet "dialog-component.js" label="component.js"}}
+{{/docs-demo}}
