--- tests/dummy/app/templates/docs/components/nucleus-button/controller.js
@@ -1,14 +1,0 @@
-// BEGIN-SNIPPET button-loading.js
-import Controller from '@ember/controller';
-import RSVP from '@ember/RSVP';
-export default Controller.extend({
-  actions: {
-    download(value) {
-      return new RSVP.Promise(function(resolve, reject) {
-        resolve(value);
-        reject();
-      });
-    }
-  }
-});
-// END-SNIPPET
