--- tests/dummy/app/templates/docs/components/nucleus-modal/demo-1/component.js
@@ -1,13 +1,0 @@
-import Component from '@ember/component';
-
-export default Component.extend({
-
-  actions: {
-    // BEGIN-SNIPPET modal-demo-simple.js
-    toggleIsShowing() {
-      this.toggleProperty('isShowing');
-    }
-    // END-SNIPPET
-  }
-
-});
